Fix [#6] 카테고리 생성, 삭제 ModalViewController로 이동
여러 도메인이 엮이므로 ModalViewController로 이동시켰습니다.
